fix(ci): allow auto-label to work for dependabot PRs (#132)
Currently, the labeler action relies on a GitHub token in a secret to
run. As Dependabot is not an org member, actions triggered from its pull
requests following [this change in
2021](https://github.com/dependabot/dependabot-core/issues/3253#issuecomment-852541544)
cannot access them, or `GITHUB_TOKEN`.

To resolve this while reducing risk of secrets exposure:
- change the trigger from `pull_request` to `pull_request_target`
- causes the labeling workflow to run in the base branch of the PR (i.e.
`main`) rather than in the PR branch
  - gives the job access to `GITHUB_TOKEN`
- change the job to use `GITHUB_TOKEN` instead of the floxbot token
- minimize the scope of `GITHUB_TOKEN` to only the ability to change PRs
- restrict execution of the labeling job to org members or dependabot